requirements.txt
adafruit-circuitpython-neopixel
adafruit-circuitpython-pypixelbuf
numpy
pygame
asyncio